refactor deployment settings

+++ deploy.sh
@@ -1,4 +1,11 @@
-cp .env.example .env
+# check if .env exists, if not copy from .env.example
+if [ ! -f .env ]; then
+    echo "Copying .env.example to .env"
+    cp .env.example .env
+else
+    echo ".env already exists, skipping copy"
+fi
+
 docker compose build
 
 docker run --rm -it --volume $(pwd):/app resepin-be composer install --no-dev --optimize-autoloader
